posix: threads_base: complete option group
Ensure that the threads_base testsuite is compatible with Linux
and that it is complete.

Update docs to reflect that.

pthread_atfork() now registers and succeeds rather than failing with
ENOSYS: fork() does not exist, so the registered handlers can never
run, and ENOSYS is not among the errors POSIX permits here.

+++ include/zephyr/posix/sys/_pthreadtypes.h
@@ -67,7 +67,7 @@ typedef struct {
 #endif
 
 #if !(defined(_PTHREAD_KEY_T_DECLARED) && defined(__pthread_key_t_defined)) || defined(__DOXYGEN__)
-typedef uint32_t pthread_key_t;
+typedef uintptr_t pthread_key_t;
 #define _PTHREAD_KEY_T_DECLARED
 #define __pthread_key_t_defined
 #endif
